automatically lint source using cpplint
This will automatically lint the source code using cpplint
in order to make sure the source code conforms with Google's
C++ style guide. One should note that if there are any issues
that the linter finds, this will not cause the job to fail.
This was choosen because there are many statements that
cause errors with the linter. Further commits should try to reduce
the number of errors by fixing true positives or by filtering out
false positives.

closes #913